Moved JsonApiSpec to the api package.  Added tests for Relationships
